trying to sort out a way of preventing the user submitting a form with empty fields
